Clean wallet.Wallet interface with functional options
Put the entry associated actions into EntriesService interface and use Wallet.Entries(options  ...Options)
method to access the service. The reason why we put it into a service is that we can add additional
info in concrete implementation. With the functional option, it provides the ability to
access the entries through the same method for different kinds of wallet. For example,
1. Generate addresses for deterministic/collection/xpub wallet:
  addrs, err := Wallet.Entries().GenerateAddresses(1)
2. Generate addresses on external chain for bip44 wallet of specific account:
  addrs, err := Wallet.Entries(Account(0)).GenerateAddresses(1)
3. Generate addresses on change chain for bip44 wallet of specific account:
  addrs, err := Wallet.Entries(Account(0), Change(true))

If no options are provided for bip44 wallet, all actions will be applied to the default account of index 0
and the external chain.