Replace the EntriesService with flat methods
It is not intuitive for users to call EntriesService to generate addresses and search addresses of a wallet.
With the flat methods and functional options, users can call wallet.GenerateAddresses(5) just like before.
Except for bip44 wallet, if no options are given, all behaviours will be applied to the external chain of
account 0.