Fix bip44 wallet newAddress error from existing wallet
The node reported a panic error when handling the newAddress API request for an existing bip44 wallet. It panics because the bip44 chains do not have the address decoder recovered when converting from a readable wallet. We can fix it by init the address decoder, however that may not be a good practise for the wallet API. As users can easily forget the init of chain decoder. Instead we should force user to pass the decoder as a param when call the `newAddresses` method, which is what we have done in this commit.