Fix wallet balance not showing - variable shadowing bug
The GetWalletBalance function had a variable shadowing issue where the
inner 'addrs' variable was shadowing the outer one, causing the address
balance map to be created from an empty slice. This resulted in the API
returning empty addresses object, making all balances show as 0.

Changed line 60 from := to = to assign to the outer addrs variable.