Fix flaky TestPexAddPeers test
The test was flaky because AddPeers shuffles addresses before adding
them when capacity is limited. The 'almost_full' test case was checking
for specific peers to be added, but due to the shuffle, different peers
could be added on each run.

Fixed by checking the total peer count instead of specific peer
addresses when capacity is limited.